Refactored to be more OOJS

+++ db/migrate/20130610173018_create_games.rb
@@ -2,7 +2,7 @@ class CreateGames < ActiveRecord::Migration
   def change
     create_table :games do |t|
       t.string :winner, :limit => 3
-      t.integer :duration
+      t.float :duration
       t.timestamps
     end
   end
